mod all_result -> nfail

+++ test/common_test/cmd_test.h
@@ -1,8 +1,8 @@
 #ifndef _CMD_TEST_H_
 #define _CMD_TEST_H_
 
-extern int8_t msh_echo_test(int8_t d);
-extern int8_t msh_exit_test(int8_t d);
-extern int8_t msh_help_test(int8_t d);
+extern uint64_t msh_echo_test(int8_t d);
+extern uint64_t msh_exit_test(int8_t d);
+extern uint64_t msh_help_test(int8_t d);
 
 #endif
